+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to
 #### Breaking Changes
 #### Added
 - syscall/samples: add `bpf/bpf_helpers.c`;
-- bpf: add history.md, c/foo.bpf.c;
+- bpf: add history.md, c/foo.bpf.c, kfunc/modules/`bpf_task_cwd_from_pid.c`;
 - qemu: add replay directory, symlink record.
 #### Changed
 #### Deprecated